Fold mount_snapshot.sh into make_snapshot.sh
Move mounting into the make_snapshot script, since that was
the only user.
Add an optional argument to make_snapshot to make it skip fsck.
Have archiveloop fsck the disk images first, then take a snapshot
of the fsck'd image using the new argument, so it does one less
fsck on startup, saving a few seconds.