setup: don't hardcode /dev/sda
DATA_DRIVE was documented as having to be set to the drive to use, but
the setup script was still ignoring its actual value and using /dev/sda
instead. Fix that and actually use the value of DATA_DRIVE, so it can
be used to specify other devices like NVME drives.