Work around Firefox issue
Firefox is throwing exceptions when accessing localstorage, so just
catch the exception and ignore it. This prevents TeslaUSB from saving
some of its settings (like the sort order of events in the dropdown,
and whether it should default to the traditional or alternative UI),
but at least the page loads again.