www: fix short sentry event
When there is a new Sentry event detected while the car is still
recording video for the previous Sentry events, it can create a 1-minute
event whose event time preceeds the recordings for the event.
In this case the car's viewer (and, prior to this change, the TeslaUSB
web viewer) would not show the event marker on the progress slider.
With this change, TeslaUSB will check if the event time from event.json
precedes the event recording timestamp, and if so it will add the
previous minute's recordings from RecentClips.